Adding BaseClass.java
