+++ src/test/java/pages/GoogleSearchResultsPage.java
@@ -1,14 +1,13 @@
 package pages;
 
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindAll;
 import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.PageFactory;
+
 import java.util.List;
 
-public class GoogleSearchResultsPage extends AbstractPage{
+public class GoogleSearchResultsPage extends AbstractPage {
 
     @FindAll({@FindBy(xpath = "//cite[contains(text(), 'wikipedia')]")})
     private List<WebElement> generalSearchResults;
@@ -18,7 +17,6 @@ public class GoogleSearchResultsPage extends AbstractPage{
     }
 
     public int countSearchResults() {
-        waitForPageLoad(driver);
         waitForPageLoad(driver);
         return generalSearchResults.size();
     }
